requirements.txt
Django>=4.2
mysqlclient>=2.2
openai>=1.3
httpx>=0.26
python-dotenv
